chore: maain types file renamed
